use new api-key header in order to fix download_metadata_zip wich broke after v11.0.3
Beginning Nudgis 11.0.3 for consistency we replaced the api endpoint for downloading metadata zip files by a 302 redirection, which was not supported properly by this reference client implementation (the api_key query string is not passed to the redirect URL). Coincidentally, v11.0.0 introduced support for passing the api-key header, which does not get dropped after a 302. This also fixes the download of zip files of slides.